+++ src/tests/gxfiletests.cpp
@@ -74,19 +74,19 @@ TEST_CASE( "Test making a good explanatory text function" )
 
    char explTxt1[256] = "'Test 1' and \"Test 2\"\t";
    std::string expectedExplTxt1 { "'Test 1' and 'Test 2'?" };
-   REQUIRE_EQ( expectedExplTxt1.length(), MakeGoodExplText( explTxt1 ) );
+   REQUIRE_EQ( static_cast<int>(expectedExplTxt1.length()), MakeGoodExplText( explTxt1 ) );
    REQUIRE( !strcmp( expectedExplTxt1.c_str(), explTxt1 ) );
 
    char explTxt2[256] = "\"Test 1\"\n and 'Test 2'";
    std::string expectedExplTxt2 { R"("Test 1"? and "Test 2")" };
-   REQUIRE_EQ( expectedExplTxt2.length(), MakeGoodExplText( explTxt2 ) );
+   REQUIRE_EQ( static_cast<int>(expectedExplTxt2.length()), MakeGoodExplText( explTxt2 ) );
    REQUIRE( !strcmp( expectedExplTxt2.c_str(), explTxt2 ) );
 
    // Make sure special chars (e.g. scandinavian alphabet, norwegian wind park names, ...) aren't
    // treated as control chars and replaced with question mark (?)
    char explTxt3[256] = "Skellefteå";
    std::string expectedExplTxt3 { "Skellefteå"s };
-   REQUIRE_EQ( expectedExplTxt3.length(), MakeGoodExplText( explTxt3 ) );
+   REQUIRE_EQ( static_cast<int>(expectedExplTxt3.length()), MakeGoodExplText( explTxt3 ) );
    REQUIRE( !strcmp( expectedExplTxt3.c_str(), explTxt3 ) );
 }
